Catalog the per-order field copy and split the answer-trigger tests
Review on #2153: the new field's label and hint were hard-coded and the
allowlist grew to hold them, and modifier-resolve.test.ts passed 750
lines with a three-case test sharing one failure. Move the copy into
modifiers.json behind t(), ratcheting the allowlist back to 20. Move
the answer-trigger tests and their setup into
modifier-resolve/answers.test.ts (the file lands at 358 lines) with
each cap scenario registered as its own test. Drop buyerVisits and
one-penny cases that duplicated modifier-resolve/counting.test.ts, and
fold the two unique buyerVisits cases in there.
