+++ src/locales/en/modifiers.json
@@ -52,6 +52,8 @@
   "modifiers.name_label": "Modifier Name",
   "modifiers.scope.none": "Nothing available to link yet.",
   "modifiers.scope.save": "Save scope",
+  "modifiers.max_per_order.label": "Max times per order (optional)",
+  "modifiers.max_per_order.hint": "Question-answer modifiers only: the most times this charge can apply within one booking. Set 1 to charge a delivery fee once per booking. Leave blank to charge for every ticket. Ignored for other triggers.",
   "modifiers.err_child_only_addon": "''{name}'' is an opt-in add-on reachable only through a required child listing, so no booking page would offer it — scope it to the parent (or another bookable listing), or remove it as a child, before saving.",
   "modifiers.answers.hint": "This modifier applies when a buyer selects any of the ticked answers.",
   "modifiers.answers.none": "No question answers to link yet.",
