+++ test/scripts/i18n-coverage.test.ts
@@ -86,7 +86,7 @@ const LEFTOVER_ALLOWLIST = new Map<string, number>([
   ["ui/templates/email/defaults.ts", 12],
   ["ui/templates/fields/add-attendee.ts", 3],
   ["ui/templates/fields/listing.ts", 4],
-  ["ui/templates/fields/modifier.ts", 22],
+  ["ui/templates/fields/modifier.ts", 20],
   ["ui/templates/fields/ticket.ts", 5],
 ]);
